@@ -5,12 +5,12 @@
   "dependencies": {
     "@turf/boolean-point-in-polygon": "^6.5.0",
     "antd": "^4.18.4",
-    "leaflet": "^1.7.1",
+    "leaflet": "^1.8.0",
     "leaflet-routing-machine": "^3.2.12",
     "proj4": "^2.7.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-leaflet": "^3.2.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.0.0",
     "react-leaflet-markercluster": "^3.0.0-rc1",
     "react-scripts": "5.0.0"
   },
